project/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
idna==2.9
mysqlclient @ file:///E:/pyProjects/django/neftehimik/project/mysqlclient-1.4.6-cp38-cp38-win32.whl
Pillow==7.1.2
pyTelegramBotAPI==3.7.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
